Add "powersave" mode to the OLED display.
This puts the OLED into powersave mode after 5 seconds. Powersave mode
means the display is turned off, but the contents remain in RAM.
The next time the layer change button is pressed, the powersave mode
is turned off, thus displaying the current layer. Pressing the layer
change button while the OLED is on changes the current layer. Then it
goes back into powersave mode after 5 seconds of no layer changes.

Pretty nice! I did this because the OLED was very bright and a somewhat
annoying presence in my peripheral vision when it was turned on all the
time.

TODO: Add a compile time option (with a #define ?) to turn off this
functionality. Alternatively it could be made into a runtime option
bound to a key.